SCMBOB-74  improved error handling on db create
improved the error handling in Import-ScDatabases when the database does
not exists and must be created. The error handling is implmented the same
as for backup restore: when an expection happens, an error the scripts
tries to get the SqlException and prints this error.
additionaly DBUtils was changed to not create the directory path, when a
path is provided for Create-Database