SCMBOB-591 refactored Set-ScSerializationReference
Since we now must set the serialization reference on the deployment system too, some changes were needed:
- Support for another template with the parameter SerializationTemplateKey
- Support for providing the exact path with param SerializationPath

The old code was way to complicated, trying to update the file when it was already here. Now we simply delete the old file a create a new one. Also we don't use xpath to get the lcoation but simply use the placeholder [BobSerializationPath]